# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BuildingStatus = typeof BuildingStatus[keyof typeof BuildingStatus];


export const BuildingStatus = {
  green: 'green',
  amber: 'amber',
  red: 'red',
} as const;

export type BuildingCategory = typeof BuildingCategory[keyof typeof BuildingCategory];


export const BuildingCategory = {
  hospital: 'hospital',
  government: 'government',
  library: 'library',
} as const;

export type BuildingCoordinates = {
  lat: number;
  lng: number;
};

export interface Building {
  id: string;
  name: string;
  address: string;
  builder: string;
  rating: number;
  status: BuildingStatus;
  category?: BuildingCategory;
  lastAudit: string;
  accessibleFeatures: string[];
  coordinates: BuildingCoordinates;
}

export type ComplaintStatus = typeof ComplaintStatus[keyof typeof ComplaintStatus];


export const ComplaintStatus = {
  Submitted: 'Submitted',
  Assigned: 'Assigned',
  In_Progress: 'In Progress',
  Resolved: 'Resolved',
  Dismissed: 'Dismissed',
} as const;

export interface Complaint {
  id: string;
  buildingId: string;
  buildingName: string;
  category: string;
  details: string;
  status: ComplaintStatus;
  officer?: string;
  dismissReason?: string;
  filedBy: string;
  submittedAt: string;
}

export interface ComplaintInput {
  /** @minLength 1 */
  buildingId: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  details: string;
  /** @minLength 1 */
  filedBy: string;
}

export type ComplaintStatusUpdateStatus = typeof ComplaintStatusUpdateStatus[keyof typeof ComplaintStatusUpdateStatus];


export const ComplaintStatusUpdateStatus = {
  Resolved: 'Resolved',
  Dismissed: 'Dismissed',
  In_Progress: 'In Progress',
} as const;

export interface ComplaintStatusUpdate {
  status: ComplaintStatusUpdateStatus;
  dismissReason?: string;
}

export interface Ngo {
  id: string;
  name: string;
  type: string;
  focus: string;
  address: string;
  tasks: string[];
}

export interface VolunteerBooking {
  id: string;
  ngoName: string;
  ngoType: string;
  date: string;
  task: string;
  occasion?: string;
}

export interface VolunteerBookingInput {
  ngoName: string;
  ngoType: string;
  date: string;
  task: string;
  occasion?: string;
}

export interface SafeSpot {
  id: string;
  name: string;
  buildingId: string;
  note: string;
}

export interface SafeSpotInput {
  name: string;
  buildingId: string;
  note: string;
}

export type GapSeverity = typeof GapSeverity[keyof typeof GapSeverity];


export const GapSeverity = {
  critical: 'critical',
  moderate: 'moderate',
  minor: 'minor',
} as const;

export interface Gap {
  id: string;
  title: string;
  severity: GapSeverity;
  reference: string;
  recommendation: string;
}

export interface ComplianceReport {
  score: number;
  rating: number;
  summary: string;
  checkedAt: string;
  gaps: Gap[];
}

export type AuditStatus = typeof AuditStatus[keyof typeof AuditStatus];


export const AuditStatus = {
  verified: 'verified',
  pending: 'pending',
} as const;

export interface Audit {
  id: string;
  auditorName: string;
  submittedAt: string;
  status: AuditStatus;
  summary: string;
}

export type BuildingDetailWayfindingItemStatus = typeof BuildingDetailWayfindingItemStatus[keyof typeof BuildingDetailWayfindingItemStatus];


export const BuildingDetailWayfindingItemStatus = {
  open: 'open',
  limited: 'limited',
  closed: 'closed',
} as const;

export type BuildingDetailWayfindingItem = {
  id: string;
  label: string;
  type: string;
  status: BuildingDetailWayfindingItemStatus;
  x: number;
  y: number;
  note: string;
};

export type BuildingDetail = Building & {
  report: ComplianceReport;
  auditor: string;
  audit: Audit;
  wayfinding: BuildingDetailWayfindingItem[];
};

export interface DashboardSummary {
  buildings: number;
  verified: number;
  openGaps: number;
  averageRating: number;
  updatedAt: string;
}

export interface ComplianceInput {
  /** @minLength 1 */
  builderName: string;
  /** @minLength 1 */
  buildingName: string;
  rampSlope: number;
  doorWidth: number;
  liftAvailable: boolean;
  accessibleRestrooms: boolean;
  tactilePath: boolean;
  blueprintName?: string;
}

export interface AuditInput {
  buildingId: string;
  /** @minLength 1 */
  auditorName: string;
  /** @minLength 1 */
  summary: string;
}

export type ListBuildingsParams = {
query?: string;
status?: ListBuildingsStatus;
};

export type ListBuildingsStatus = typeof ListBuildingsStatus[keyof typeof ListBuildingsStatus];


export const ListBuildingsStatus = {
  all: 'all',
  green: 'green',
  amber: 'amber',
  red: 'red',
} as const;

export type CreateBuddyRequest200 = {
  success?: boolean;
};

export type GetUserStrikes200 = {
  strikes?: number;
};