# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: buildings
    description: Public building directory and accessibility data
  - name: compliance
    description: Blueprint compliance checking and auditor submissions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /buildings:
    get:
      operationId: listBuildings
      tags: [buildings]
      summary: List accessible buildings
      parameters:
        - in: query
          name: query
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [all, green, amber, red]
      responses:
        "200":
          description: Building directory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
  /buildings/{id}:
    get:
      operationId: getBuilding
      tags: [buildings]
      summary: Get a building and its transparency record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Building detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingDetail"
        "404":
          description: Building not found
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [buildings]
      summary: Get public dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /compliance/check:
    post:
      operationId: runComplianceCheck
      tags: [compliance]
      summary: Run the accessibility rule engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceInput"
      responses:
        "200":
          description: Compliance gap report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReport"
  /audits:
    post:
      operationId: submitAudit
      tags: [compliance]
      summary: Submit a field audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditInput"
      responses:
        "201":
          description: Submitted audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audit"
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: submitComplaint
      tags: [complaints]
      summary: Submit a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Submitted complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    patch:
      operationId: updateComplaintStatus
      tags: [complaints]
      summary: Update complaint status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintStatusUpdate"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /ngos:
    get:
      operationId: listNGOs
      tags: [volunteering]
      summary: List registered NGOs
      responses:
        "200":
          description: List of NGOs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NGO"
  /volunteer-bookings:
    get:
      operationId: listVolunteerBookings
      tags: [volunteering]
      summary: List volunteer bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VolunteerBooking"
    post:
      operationId: createVolunteerBooking
      tags: [volunteering]
      summary: Book a volunteer slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerBookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerBooking"
  /safe-spots:
    get:
      operationId: listSafeSpots
      tags: [safe-spots]
      summary: List safe spots
      responses:
        "200":
          description: List of safe spots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SafeSpot"
    post:
      operationId: createSafeSpot
      tags: [safe-spots]
      summary: Create a safe spot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SafeSpotInput"
      responses:
        "201":
          description: Created safe spot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeSpot"
  /buddy-requests:
    post:
      operationId: createBuddyRequest
      tags: [buddy]
      summary: Broadcast a buddy request
      responses:
        "200":
          description: Buddy request sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
  /users/{name}/strikes:
    get:
      operationId: getUserStrikes
      tags: [users]
      summary: Get fake strike count for a user
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User strikes count
          content:
            application/json:
              schema:
                type: object
                properties:
                  strikes:
                    type: number
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Building:
      type: object
      required: [id, name, address, builder, rating, status, lastAudit, accessibleFeatures, coordinates]
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        builder: { type: string }
        rating: { type: number }
        status:
          type: string
          enum: [green, amber, red]
        category:
          type: string
          enum: [hospital, government, library]
        lastAudit: { type: string }
        accessibleFeatures:
          type: array
          items: { type: string }
        coordinates:
          type: object
          required: [lat, lng]
          properties:
            lat: { type: number }
            lng: { type: number }
    Complaint:
      type: object
      required: [id, buildingId, buildingName, category, details, status, filedBy, submittedAt]
      properties:
        id: { type: string }
        buildingId: { type: string }
        buildingName: { type: string }
        category: { type: string }
        details: { type: string }
        status:
          type: string
          enum: [Submitted, Assigned, In Progress, Resolved, Dismissed]
        officer: { type: string }
        dismissReason: { type: string }
        filedBy: { type: string }
        submittedAt: { type: string }
    ComplaintInput:
      type: object
      required: [buildingId, category, details, filedBy]
      properties:
        buildingId: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        details: { type: string, minLength: 1 }
        filedBy: { type: string, minLength: 1 }
    ComplaintStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Resolved, Dismissed, In Progress]
        dismissReason:
          type: string
    NGO:
      type: object
      required: [id, name, type, focus, address, tasks]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        focus: { type: string }
        address: { type: string }
        tasks:
          type: array
          items: { type: string }
    VolunteerBooking:
      type: object
      required: [id, ngoName, ngoType, date, task]
      properties:
        id: { type: string }
        ngoName: { type: string }
        ngoType: { type: string }
        date: { type: string }
        task: { type: string }
        occasion: { type: string }
    VolunteerBookingInput:
      type: object
      required: [ngoName, ngoType, date, task]
      properties:
        ngoName: { type: string }
        ngoType: { type: string }
        date: { type: string }
        task: { type: string }
        occasion: { type: string }
    SafeSpot:
      type: object
      required: [id, name, buildingId, note]
      properties:
        id: { type: string }
        name: { type: string }
        buildingId: { type: string }
        note: { type: string }
    SafeSpotInput:
      type: object
      required: [name, buildingId, note]
      properties:
        name: { type: string }
        buildingId: { type: string }
        note: { type: string }
    Gap:
      type: object
      required: [id, title, severity, reference, recommendation]
      properties:
        id: { type: string }
        title: { type: string }
        severity:
          type: string
          enum: [critical, moderate, minor]
        reference: { type: string }
        recommendation: { type: string }
    ComplianceReport:
      type: object
      required: [score, rating, summary, gaps, checkedAt]
      properties:
        score: { type: number }
        rating: { type: number }
        summary: { type: string }
        checkedAt: { type: string }
        gaps:
          type: array
          items: { $ref: "#/components/schemas/Gap" }
    Audit:
      type: object
      required: [id, auditorName, submittedAt, status, summary]
      properties:
        id: { type: string }
        auditorName: { type: string }
        submittedAt: { type: string }
        status:
          type: string
          enum: [verified, pending]
        summary: { type: string }
    BuildingDetail:
      allOf:
        - $ref: "#/components/schemas/Building"
        - type: object
          required: [report, auditor, audit, wayfinding]
          properties:
            report: { $ref: "#/components/schemas/ComplianceReport" }
            auditor: { type: string }
            audit: { $ref: "#/components/schemas/Audit" }
            wayfinding:
              type: array
              items:
                type: object
                required: [id, label, type, status, x, y, note]
                properties:
                  id: { type: string }
                  label: { type: string }
                  type: { type: string }
                  status:
                    type: string
                    enum: [open, limited, closed]
                  x: { type: number }
                  y: { type: number }
                  note: { type: string }
    DashboardSummary:
      type: object
      required: [buildings, verified, openGaps, averageRating, updatedAt]
      properties:
        buildings: { type: number }
        verified: { type: number }
        openGaps: { type: number }
        averageRating: { type: number }
        updatedAt: { type: string }
    ComplianceInput:
      type: object
      required: [builderName, buildingName, rampSlope, doorWidth, liftAvailable, accessibleRestrooms, tactilePath]
      properties:
        builderName: { type: string, minLength: 1 }
        buildingName: { type: string, minLength: 1 }
        rampSlope: { type: number }
        doorWidth: { type: number }
        liftAvailable: { type: boolean }
        accessibleRestrooms: { type: boolean }
        tactilePath: { type: boolean }
        blueprintName: { type: string }
    AuditInput:
      type: object
      required: [buildingId, auditorName, summary]
      properties:
        buildingId: { type: string }
        auditorName: { type: string, minLength: 1 }
        summary: { type: string, minLength: 1 }