# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatus } from './complaintStatus';

export interface Complaint {
  id: string;
  buildingId: string;
  buildingName: string;
  category: string;
  details: string;
  status: ComplaintStatus;
  officer?: string;
  dismissReason?: string;
  filedBy: string;
  submittedAt: string;
}