# lib/api-zod/src/generated/types/complaintStatusUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatusUpdateStatus } from './complaintStatusUpdateStatus';

export interface ComplaintStatusUpdate {
  status: ComplaintStatusUpdateStatus;
  dismissReason?: string;
}