# lib/api-zod/src/generated/types/volunteerBooking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VolunteerBooking {
  id: string;
  ngoName: string;
  ngoType: string;
  date: string;
  task: string;
  occasion?: string;
}